Fixed detection of AVX512 intrinsics for -O3 builds (i.e. release)
- Moved SSE detection before setting the optimization flags
- AVX512 detection code was optimized out so test succeeded even if AVX512 was not supported

- Fix return type of generator function

+++ src/gui/performance_plot.cpp
@@ -16,7 +16,7 @@ PerformancePlot::PerformancePlot(Settings* p_glob_settings, SpectrumAdapter* p_s
 
   rnti_x_axis = std::vector<double>(65536);
   int x = 0;
-  std::generate(rnti_x_axis.begin(), rnti_x_axis.end(), [&]{ return x++; });
+  std::generate(rnti_x_axis.begin(), rnti_x_axis.end(), [&]{ return static_cast<double>(x++); });
 
   //Generate Window PLOT_A_WINDOW:
   plot_a_window = new QWidget();
